# Limit YJIT stats to the measured benchmark code