--- a/harness/harness-common.rb
+++ b/harness/harness-common.rb
@@ -101,6 +101,9 @@ def return_results(warmup_iterations, bench_iterations)
     "bench" => bench_iterations,
   }
 
+  # Collect yjit stats before loading any additional code.
+  yjit_stats = RubyVM::YJIT.runtime_stats if defined?(RubyVM::YJIT) && RubyVM::YJIT.enabled?
+
   # Collect our own peak mem usage as soon as reasonable after finishing the last iteration.
   rss = get_rss
   yjit_bench_results["rss"] = rss
@@ -108,8 +111,7 @@ def return_results(warmup_iterations, bench_iterations)
     yjit_bench_results["maxrss"] = maxrss
   end
 
-  if defined?(RubyVM::YJIT) && RubyVM::YJIT.enabled?
-    yjit_stats = RubyVM::YJIT.runtime_stats
+  if yjit_stats
     yjit_bench_results["yjit_stats"] = yjit_stats
 
     formatted_stats = proc { |key| "%10s" % yjit_stats[key].to_s.reverse.scan(/\d{1,3}/).join(",").reverse }