Reset stats at the beginning of the benchmark
and collect stats before loading any additional code.

You can see a difference with the invalidation count:

before:
```
$  MIN_BENCH_ITRS=1 WARMUP_ITRS=1 MIN_BENCH_TIME=0 RBENV_VERSION=ruby-master YJIT_BENCH_STATS=invalidation_count ruby --yjit-stats=quiet benchmarks/sequel/benchmark.rb
ruby 3.4.0dev (2024-12-04T21:34:16Z master 1c4dbb133e) +YJIT stats +PRISM [arm64-darwin23]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
itr:   time invalidation_count
 #1:   31ms                  0
 #2:   30ms                  0
invalidation_count:         68
```
after:
```
$  MIN_BENCH_ITRS=1 WARMUP_ITRS=1 MIN_BENCH_TIME=0 RBENV_VERSION=ruby-master YJIT_BENCH_STATS=invalidation_count ruby --yjit-stats=quiet benchmarks/sequel/benchmark.rb
ruby 3.4.0dev (2024-12-04T21:34:16Z master 1c4dbb133e) +YJIT stats +PRISM [arm64-darwin23]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
itr:   time invalidation_count
 #1:   31ms                  0
 #2:   30ms                  0
invalidation_count:          0
```