@maximecb 2024-12-09
@rwstauner for some reason the test is failing right at the end:

```
Building native extensions. This could take a while...
/opt/hostedtoolcache/Ruby/3.3.6/x64/lib/ruby/3.3.0/rubygems/ext/builder.rb:125:in `run': ERROR: Failed to build gem native extension. (Gem::Ext::BuildError)
```

Not immediately clear which gem is failing to build and why this happens so late in the process 🤔