@rwstauner 2024-12-09
The test started failing because the GitHub runner upgraded:
https://github.com/Shopify/yjit-bench/pull/353